Add ability to request multiple certificates (#4)
Allow a client to request multiple certificates from the certificate provider and, optionally, allow a chain to be provided to the clients. A server can build up multiple requests by using the add_request_server_cert method and then calling request_server_certs when ready eg

  @when('certificates.available')
  def request_cert(tls):
    tls.add_request_server_cert('admin.keystone.openstack.local', ['10.0.0.23', 'alt.openstack.local])
    tls.add_request_server_cert('internal.keystone.openstack.local', [])
    tls.request_server_certs()

Under the covers a new key 'cert_requests' is used on the relation. This is set to a json dump of a dictionary representing the requests.

The certificates provider then uses add_server_cert to build up the response before sending the response with set_server_multicerts eg

  tls.add_server_cert(unit_name, 'admin.keystone.openstack.local', 'CERT1', 'KEY1')
  tls.add_server_cert(unit_name, ''internal.keystone.openstack.local', 'CERT2', 'KEY2')
  tls.set_server_multicerts(unit_name)