Fix IndexError due to handler dispatch order uncertainty
We need to be a bit more defensive in how we treat data because charm
handlers might be run before our `broken` handler.  The real solution
for this is to give the `broken` handler some measure of priority, or
use clear triggers, but neither of those are available yet.  Also,
coding errors on the charm side could still potentially lead to methods
being called when they shouldn't, so we should be more defensive anyway.