Fix cross-model relations
With cross-model relations (CMR), the "unit name" visible on the
offering side of the relation is a UUID which doesn't match with the
unit's own view of its unit name.  Thus, the unit cannot find the
responses to its cert requests, as they are keyed by the UUID rather
than the unit name.  By explicitly publishing the unit name over the
relation, it ensures that the provider and requirer will use the same
key.

We use the unit name rather than a UUID or nonce to ensure that non-CMR
deployments are not broken upon upgrade.

Fixes #14