Add `connected` state to improve downstream charm messaging
Some charms have a distinction between something being alive in
the other end of a relation (`connected`) and the other end
actually being ready for prime time (`available`).

We don't have that distinction in the top level flags in this
interface but let's set both states to improve the messaging of
said charms.

Fixes #23